23/7/2026 19:31 ++_(janitoring an ancient code base, trying to re grasp it )

+++ sharedStructs.h
@@ -26,10 +26,10 @@
 #pragma pack(push, 1)
 struct KnockPacket {
     char magic[128];
-    int tcpReturn;
+    uint16_t tcpReturn;
     std::string publicKey;
     uint32_t publicKeyLen;
-    int filePort;
+    uint16_t filePort;
     int latencyOfConnection;
 };
 #pragma pack(pop)
